+++ src/app/layout.tsx
@@ -1,4 +1,4 @@
-import { Header } from './_components/Header';
+import { Header } from '../components/Header';
 import './globals.css';
 import { Inter } from 'next/font/google';
 import type { Metadata } from 'next';
